chore: Migrate aetherResolver & API codegen to zod
