+++ packages/@aetherspace/schemas/aetherGraphSchema.ts
@@ -73,7 +73,6 @@ const aetherSchemaDefinitions = (aetherSchema: ResolverSchemaType, prefix = 'typ
     AetherObject: createDefinition('Schema'),
     // -- Arraylikes --
     AetherArray: createDefinition('Array'),
-    AetherCollection: createDefinition('Array'),
   })
   // Transform into usable graphql definitions (TODO: Figure out why aetherSchema can be undefined, yet graphql doesn't break?)
   const schemaDef = `
