+++ packages/@aetherspace/schemas/aetherSchemas.ts
@@ -1,281 +1,427 @@
-import * as ss from 'superstruct'
-import type {
-  ObjectSchema,
-  ObjectType,
-  StructSchema,
-  Simplify,
-  Optionalize,
-} from 'superstruct/lib/utils'
+import { z } from 'zod'
+import type { JSONSchema7 } from 'json-schema'
+import zodToJsonSchema from 'zod-to-json-schema'
 
 /* --- Types ----------------------------------------------------------------------------------- */
 
-export type AetherSchemaType<T = any, S = any> = ss.Struct<T, S>
-export type Infer<T extends AetherSchemaType> = T['TYPE'] // Same as superstruct's Infer type
-export type Describe<T> = ss.Struct<T, StructSchema<T>> // Same as superstruct's Describe type
+const SCHEMA_MAP = Object.freeze({
+  boolean: 'AetherBoolean',
+  number: 'AetherNumber',
+  string: 'AetherString',
+  date: 'AetherDate',
+  object: 'AetherSchema',
+  array: 'AetherArray',
+  color: 'AetherColor',
+  id: 'AetherId',
+})
 
-export type { Simplify, Optionalize, ObjectSchema, ObjectType, StructSchema }
-
-// -i- Use to extend individual schema entries
-export type SchemaEntry<T extends AetherSchemaType> = {
-  type: string
-  aetherType: string
-  schema: T['schema'] | null
-  optional?: boolean
-  nullable?: boolean
-  default?: unknown
-  example?: T['TYPE']
+export type AetherSchemaType<T = any> = {
+  type: keyof typeof SCHEMA_MAP
+  aetherType: typeof SCHEMA_MAP[keyof typeof SCHEMA_MAP]
   description?: string
+  defaultValue?: T
   schemaName?: string
+  isOptional?: boolean
+  isNullable?: boolean
+  schema: AetherSchemaType | Record<string, any>
 }
 
 export type SchemaPluginMap = {
   // -- Primitives --
-  AetherString: (name: string, schema: SchemaEntry<ReturnType<typeof AetherString>>) => unknown
-  AetherNumber: (name: string, schema: SchemaEntry<ReturnType<typeof AetherNumber>>) => unknown
-  AetherBoolean: (name: string, schema: SchemaEntry<ReturnType<typeof AetherBoolean>>) => unknown
+  AetherString: (name: string, schema: AetherSchemaType) => unknown
+  AetherNumber: (name: string, schema: AetherSchemaType) => unknown
+  AetherBoolean: (name: string, schema: AetherSchemaType) => unknown
   // -- Single values --
-  AetherId: (name: string, schema: SchemaEntry<ReturnType<typeof AetherID>>) => unknown
-  AetherColor: (name: string, schema: SchemaEntry<ReturnType<typeof AetherColor>>) => unknown
-  AetherDate: (name: string, schema: SchemaEntry<ReturnType<typeof AetherDate>>) => unknown
-  AetherEnum: (name: string, schema: SchemaEntry<ReturnType<typeof AetherEnum>>) => unknown
+  AetherId: (name: string, schema: AetherSchemaType) => unknown
+  AetherColor: (name: string, schema: AetherSchemaType) => unknown
+  AetherDate: (name: string, schema: AetherSchemaType) => unknown
+  AetherEnum: (name: string, schema: AetherSchemaType) => unknown
   // -- Objectlikes --
-  AetherSchema: (name: string, schema: SchemaEntry<ReturnType<typeof AetherSchema>>) => unknown
-  AetherObject: (name: string, schema: SchemaEntry<ReturnType<typeof AetherSchema>>) => unknown
+  AetherSchema: (name: string, schema: AetherSchemaType) => unknown
+  AetherObject: (name: string, schema: AetherSchemaType) => unknown
   // -- Arraylikes --
-  AetherArray: (name: string, schema: SchemaEntry<ReturnType<typeof AetherArray>>) => unknown
-  AetherCollection: (name: string, schema: SchemaEntry<ReturnType<typeof AetherCollection>>) => unknown // prettier-ignore
+  AetherArray: (name: string, schema: AetherSchemaType) => unknown
 }
 
-/* --- Helpers --------------------------------------------------------------------------------- */
+declare module 'zod' {
+  interface ZodType<Output> {
+    exampleValue?: Output
+  }
 
-const assignDescriptors = <R extends AetherSchemaType>(
-  schema: R,
-  aetherType: string,
-  schemaName?: string
-) => {
-  return Object.assign(schema, {
-    // Chain command for docs, indicating example value & description to schema property (e.g. for Storybook)
-    docs: (example, description?: string) => Object.assign(schema, { example, description }),
-    // Chain command for indicating default value + add example & description for docs
-    default: (defaultValue, description?: string, example?: any) => {
-      return Object.assign(schema, {
-        defaultValue,
-        example: example || defaultValue,
-        ...(description ? { description } : null),
-      })
-    },
-    // Assign type to schema object for later interpretation
-    aetherType,
-    // Assign schemaName (if known) to object for later interpretation
-    ...(schemaName ? { schemaName } : null),
-  })
-}
+  interface ZodString {
+    aetherType: 'AetherString'
+    example(value: string): z.ZodString
+    eg(value: string): z.ZodString
+    ex(value: string): z.ZodString
+  }
 
-const makeOptionalable = <T, S, ST extends ss.Struct<T, S>>(
-  schema: ST,
-  aetherType: string,
-  schemaName?: string
-) => {
-  return Object.assign(schema, {
-    // Chain command to indicate the field can be nullable (e.g. in GraphQL)
-    nullable: () => {
-      const nullableSchema = Object.assign(ss.nullable(schema), { isNullable: true })
-      return assignDescriptors(nullableSchema, aetherType, schemaName)
-    },
-    // Chain command to indicate the field can be nullish (= null or undefined)
-    nullish: () => {
-      const nullableSchema = Object.assign(ss.nullable(schema), { isNullable: true })
-      const nullishSchema = Object.assign(ss.optional(nullableSchema), {
-        isNullable: true,
-        isOptional: true,
-      })
-      return assignDescriptors(nullishSchema, aetherType, schemaName)
-    },
-    // Chain command to indicate the field can be omitted / undefined (e.g. in Docs)
-    optional: () => {
-      const optionalSchema = Object.assign(ss.optional(schema), { isOptional: true })
-      return assignDescriptors(optionalSchema, aetherType, schemaName)
-    },
-  })
-}
+  interface ZodNumber {
+    aetherType: 'AetherNumber'
+    example(value: number): z.ZodNumber
+    eg(value: number): z.ZodNumber
+    ex(value: number): z.ZodNumber
+  }
 
-const makeExtendable = <S extends ObjectSchema>(schema: ss.Struct<ObjectType<S>, S>) => {
-  return Object.assign(schema, {
-    // Extend the schema with additional properties
-    extend: <SE extends ObjectSchema>(newSchemaName: string, extraPropsSchema: SE) => {
-      const newSchema = extendSchema(newSchemaName, schema, extraPropsSchema)
-      return makeExtendable(newSchema)
-    },
-    // Extend the schema by omitting properties
-    omit: <K extends keyof S>(newSchemaName: string, keys: K[]) => {
-      const newSchema = omitSchemaProps(newSchemaName, schema, keys)
-      return makeExtendable(newSchema)
-    },
-    // Make the entire schema partial
-    partial: (newSchemaName: string) => {
-      return makePartialSchema(newSchemaName, schema)
-    },
-    // Create new schema with specific properties from the original
-    pick: <K extends keyof S>(newSchemaName: string, keys: K[]) => {
-      const newSchema = pickSchemaProps(newSchemaName, schema, keys)
-      return makeExtendable(newSchema)
-    },
-  })
-}
+  interface ZodBoolean {
+    aetherType: 'AetherBoolean'
+    example(value: boolean): z.ZodBoolean
+    eg(value: boolean): z.ZodBoolean
+    ex(value: boolean): z.ZodBoolean
+  }
 
-const aetherWrapper = <A extends any[], T, S>(
-  struct: (...args: A) => ss.Struct<T, S>,
-  aetherType: string
-) => {
-  return (...args: A) => {
-    // Add chain commands for extending with default or documentation descriptions
-    const schema = assignDescriptors(struct(...args), aetherType)
-    // Add chain commands for nullable & optional fields or properties
-    return makeOptionalable<T, S, typeof schema>(schema, aetherType)
+  interface ZodDate {
+    aetherType: 'AetherDate'
+    example(value: Date): z.ZodDate
+    eg(value: Date): z.ZodDate
+    ex(value: Date): z.ZodDate
+  }
+
+  interface ZodEnum<T extends [string, ...string[]]> {
+    aetherType: 'AetherEnum'
+    example(value: z.infer<z.ZodEnum<T>>): z.ZodEnum<T>
+    eg(value: z.infer<z.ZodEnum<T>>): z.ZodEnum<T>
+    ex(value: z.infer<z.ZodEnum<T>>): z.ZodEnum<T>
+  }
+
+  interface ZodTuple<T extends [] | [z.ZodTypeAny, ...z.ZodTypeAny[]]> {
+    aetherType: 'AetherTuple'
+    example(value: z.infer<z.ZodTuple<T>>): z.ZodTuple<T>
+    eg(value: z.infer<z.ZodTuple<T>>): z.ZodTuple<T>
+    ex(value: z.infer<z.ZodTuple<T>>): z.ZodTuple<T>
+  }
+
+  interface ZodUnion<T extends readonly [z.ZodTypeAny, ...z.ZodTypeAny[]]> {
+    aetherType: 'AetherUnion'
+    example(value: z.infer<z.ZodUnion<T>>): z.ZodUnion<T>
+    eg(value: z.infer<z.ZodUnion<T>>): z.ZodUnion<T>
+    ex(value: z.infer<z.ZodUnion<T>>): z.ZodUnion<T>
+  }
+
+  interface ZodArray<T extends z.ZodTypeAny> {
+    aetherType: 'AetherArray'
+    example(value: z.infer<z.ZodArray<T>>): z.ZodArray<T>
+    eg(value: z.infer<z.ZodArray<T>>): z.ZodArray<T>
+    ex(value: z.infer<z.ZodArray<T>>): z.ZodArray<T>
+  }
+
+  interface ZodObject<
+    T extends z.ZodRawShape,
+    UnknownKeys extends 'passthrough' | 'strict' | 'strip' = 'strip',
+    Catchall extends z.ZodTypeAny = z.ZodTypeAny,
+    Output = z.objectOutputType<T, Catchall>,
+    Input = z.objectInputType<T, Catchall>
+  > {
+    aetherType: 'AetherObject' | 'AetherSchema'
+    schemaName?: string
+    extendedFrom?: string
+    nameSchema(schemaName: string): z.ZodObject<T, UnknownKeys, Catchall, Output, Input>
+    extendSchema<E extends z.ZodRawShape>(
+      schemaName: string,
+      extraFields: E
+    ): z.ZodObject<
+      T & E,
+      UnknownKeys,
+      Catchall,
+      z.objectOutputType<T & E, Catchall>,
+      z.objectInputType<T & E, Catchall>
+    >
+    pickSchema<
+      P extends Parameters<z.ZodObject<T>['pick']>[0],
+      K extends Exclude<Exclude<keyof P, undefined>, symbol>
+    >(
+      schemaName: string,
+      picks: P
+    ): z.ZodObject<
+      Pick<T, K>,
+      UnknownKeys,
+      Catchall,
+      z.objectOutputType<Pick<T, K>, Catchall>,
+      z.objectInputType<Pick<T, K>, Catchall>
+    >
+    omitSchema<
+      P extends Parameters<z.ZodObject<T>['omit']>[0],
+      K extends Exclude<Exclude<keyof P, undefined>, symbol>
+    >(
+      schemaName: string,
+      picks: P
+    ): z.ZodObject<
+      Omit<T, K>,
+      UnknownKeys,
+      Catchall,
+      z.objectOutputType<Omit<T, K>, Catchall>,
+      z.objectInputType<Omit<T, K>, Catchall>
+    >
+    partialSchema(
+      schemaName: string
+    ): ReturnType<z.ZodObject<T, UnknownKeys, Catchall, Output, Input>['partial']>
+    deepPartialSchema(
+      schemaName: string
+    ): ReturnType<z.ZodObject<T, UnknownKeys, Catchall, Output, Input>['deepPartial']>
+    requiredSchema(
+      schemaName: string
+    ): ReturnType<z.ZodObject<T, UnknownKeys, Catchall, Output, Input>['required']>
+    jsonSchema(): JSONSchema7
+    introspect(): Record<string, any>
+    example(value: z.infer<z.ZodObject<T>>): z.ZodObject<T>
+    eg(value: z.infer<z.ZodObject<T>>): z.ZodObject<T>
+    ex(value: z.infer<z.ZodObject<T>>): z.ZodObject<T>
   }
 }
 
-/* --- Primitive Schema Types ------------------------------------------------------------------ */
+/* --- Primitives ------------------------------------------------------------------------------ */
 
-const AetherID = aetherWrapper(ss.string, 'AetherID')
-const AetherColor = aetherWrapper(ss.string, 'AetherColor')
-const AetherString = aetherWrapper(ss.string, 'AetherString')
-const AetherNumber = aetherWrapper(ss.number, 'AetherNumber')
-const AetherBoolean = aetherWrapper(ss.boolean, 'AetherBoolean')
-const AetherDate = aetherWrapper(ss.date, 'AetherDate')
-
-const AetherEnum = <T extends string = string>(values: readonly T[]) => {
-  const schema = assignDescriptors(ss.enums<T>(values), 'AetherEnum')
-  return makeOptionalable<T, typeof schema['schema'], typeof schema>(schema, 'AetherEnum')
-}
-
-/* --- Advanced Schema Types ------------------------------------------------------------------- */
-
-const AetherSchema = <S extends ObjectSchema>(schemaName: string, objSchema: S) => {
-  const schema = assignDescriptors(ss.object(objSchema), 'AetherSchema', schemaName)
-  const optschema = makeOptionalable<ObjectType<S>, typeof schema['schema'], typeof schema>(schema, 'AetherSchema', schemaName) // prettier-ignore
-  return makeExtendable(optschema)
-}
-
-const AetherArray = <T extends ss.Struct<any>>(Element: T) => {
-  const arraySchema = assignDescriptors(ss.array<T>(Element), 'AetherArray')
-  return makeOptionalable<ss.Infer<T>[], typeof arraySchema['schema'], typeof arraySchema>(arraySchema, 'AetherArray') // prettier-ignore
-}
-
-const AetherCollection = <S extends ObjectSchema>(schemaName: string, objSchema: S) => {
-  const entrySchema = AetherSchema(schemaName, objSchema)
-  return AetherArray(entrySchema)
-}
-
-/* --- Utilities ------------------------------------------------------------------------------- */
-
-const extendSchema = <A extends ObjectSchema, B extends ObjectSchema>(
-  schemaName: string,
-  originalSchema: ss.Struct<ObjectType<A>, A>,
-  newProperties: B
-) => {
-  // Create new schema by extending original schema with new properties using superstruct's assign() method
-  const schema = assignDescriptors(ss.assign(originalSchema, ss.object(newProperties)), 'AetherSchema', schemaName) // prettier-ignore
-  return makeOptionalable<typeof schema['TYPE'], typeof schema['schema'], typeof schema>(schema, 'AetherSchema', schemaName) // prettier-ignore
-}
-
-const omitSchemaProps = <S extends ObjectSchema, K extends keyof S>(
-  schemaName: string,
-  originalSchema: ss.Struct<ObjectType<S>, S>,
-  keys: K[]
-) => {
-  const schema = assignDescriptors(ss.omit(originalSchema, keys), 'AetherSchema', schemaName)
-  return makeOptionalable<typeof schema['TYPE'], typeof schema['schema'], typeof schema>(schema, 'AetherSchema', schemaName) // prettier-ignore
-}
-
-const makePartialSchema = <S extends ObjectSchema>(
-  schemaName: string,
-  originalSchema: ss.Struct<ObjectType<S>, S>
-) => {
-  const schema = assignDescriptors(ss.partial(originalSchema), 'AetherSchema', schemaName)
-  for (const key in schema.schema) {
-    const { aetherType, isNullable } = originalSchema.schema[key] as AetherSchemaType['schema'][string] // prettier-ignore
-    schema.schema[key] = Object.assign(schema.schema[key], { aetherType, isNullable, isOptional: true }) // prettier-ignore
+if (!z.ZodString.prototype?.aetherType) {
+  z.ZodString.prototype.aetherType = 'AetherString'
+  z.ZodString.prototype.example = function (value: string) {
+    this.exampleValue = value
+    return this
   }
-  return makeOptionalable<typeof schema['TYPE'], typeof schema['schema'], typeof schema>(schema, 'AetherSchema', schemaName) // prettier-ignore
+  z.ZodString.prototype.eg = z.ZodString.prototype.example
+  z.ZodString.prototype.ex = z.ZodString.prototype.example
 }
 
-const pickSchemaProps = <S extends ObjectSchema, K extends keyof S>(
-  schemaName: string,
-  originalSchema: ss.Struct<ObjectType<S>, S>,
-  keys: K[]
-) => {
-  const schema = assignDescriptors(ss.pick(originalSchema, keys), 'AetherSchema', schemaName)
-  for (const key in schema.schema) {
-    const { aetherType, isOptional, isNullable } = originalSchema.schema[key] as AetherSchemaType['schema'][string] // prettier-ignore
-    schema.schema[key] = Object.assign(schema.schema[key], { aetherType, isOptional, isNullable }) // prettier-ignore
+if (!z.ZodNumber.prototype.aetherType) {
+  z.ZodNumber.prototype.aetherType = 'AetherNumber'
+  z.ZodNumber.prototype.example = function (value: number) {
+    this.exampleValue = value
+    return this
   }
-  return makeOptionalable<typeof schema['TYPE'], typeof schema['schema'], typeof schema>(schema, 'AetherSchema', schemaName) // prettier-ignore
+  z.ZodNumber.prototype.eg = z.ZodNumber.prototype.example
+  z.ZodNumber.prototype.ex = z.ZodNumber.prototype.example
 }
 
-/* --- Helpers --------------------------------------------------------------------------------- */
+if (!z.ZodBoolean.prototype.aetherType) {
+  z.ZodBoolean.prototype.aetherType = 'AetherBoolean'
+  z.ZodBoolean.prototype.example = function (value: boolean) {
+    this.exampleValue = value
+    return this
+  }
+  z.ZodBoolean.prototype.eg = z.ZodBoolean.prototype.example
+  z.ZodBoolean.prototype.ex = z.ZodBoolean.prototype.example
+}
 
-export const applySchema = <S extends AetherSchemaType>(
-  obj: Record<string, unknown> | Infer<S>,
-  schema: S,
-  applyExamples = false
-): Infer<S> => {
-  // Loop over schemaConfig
-  const objWithDefaults = Object.entries(schema.schema).reduce((acc, entries) => {
-    // Extract schema entries
-    const [name, fieldConfig] = entries as [string, { defaultValue?: any; example?: any, type: string }] // prettier-ignore
-    // Detect if we should nest
-    const isObjectSchema = ['object', 'schema'].includes(fieldConfig.type)
-    // Always apply the same for nested objects
-    if (isObjectSchema) {
-      return {
-        ...acc,
-        [name]: applySchema(obj?.[name] || {}, fieldConfig as AetherSchemaType, applyExamples),
-      }
+/* --- Advanced -------------------------------------------------------------------------------- */
+
+if (!z.ZodDate.prototype.aetherType) {
+  z.ZodDate.prototype.aetherType = 'AetherDate'
+  z.ZodDate.prototype.example = function (value: Date) {
+    this.exampleValue = value
+    return this
+  }
+  z.ZodDate.prototype.eg = z.ZodDate.prototype.example
+  z.ZodDate.prototype.ex = z.ZodDate.prototype.example
+}
+
+if (!z.ZodEnum.prototype.aetherType) {
+  z.ZodEnum.prototype.aetherType = 'AetherEnum'
+  z.ZodEnum.prototype.example = function (value: string) {
+    this.exampleValue = value
+    return this
+  }
+  z.ZodEnum.prototype.eg = z.ZodEnum.prototype.example
+  z.ZodEnum.prototype.ex = z.ZodEnum.prototype.example
+}
+
+if (!z.ZodTuple.prototype.aetherType) {
+  z.ZodTuple.prototype.aetherType = 'AetherTuple'
+  z.ZodTuple.prototype.example = function (value: any[]) {
+    this.exampleValue = value
+    return this
+  }
+  z.ZodTuple.prototype.eg = z.ZodTuple.prototype.example
+  z.ZodTuple.prototype.ex = z.ZodTuple.prototype.example
+}
+
+if (!z.ZodUnion.prototype.aetherType) {
+  z.ZodUnion.prototype.aetherType = 'AetherUnion'
+  z.ZodUnion.prototype.example = function (value: any) {
+    this.exampleValue = value
+    return this
+  }
+  z.ZodUnion.prototype.eg = z.ZodUnion.prototype.example
+  z.ZodUnion.prototype.ex = z.ZodUnion.prototype.example
+}
+
+if (!z.ZodArray.prototype.aetherType) {
+  z.ZodArray.prototype.aetherType = 'AetherArray'
+  z.ZodArray.prototype.example = function (value: any[]) {
+    this.exampleValue = value
+    return this
+  }
+  z.ZodArray.prototype.eg = z.ZodArray.prototype.example
+  z.ZodArray.prototype.ex = z.ZodArray.prototype.example
+}
+
+/** --- parseType() ---------------------------------------------------------------------------- */
+/** -i- Add correct fields like type & aetherType while converting a JSON schema prop definition to aetherspace prop definition. */
+const parseType = (propDef: JSONSchema7, propSchema: Record<string, unknown> = {}) => {
+  // @ts-ignore
+  const { type: propType, enum: propEnum, description, anyOf } = propDef
+  if (Array.isArray(propType)) {
+    propSchema.isNullable = propType.includes('null')
+    const leftoverType = propType.filter((type: string) => type !== 'null')[0]
+    propSchema.aetherType = SCHEMA_MAP[leftoverType]
+    propSchema.type = leftoverType
+  } else if (propDef.format === 'date-time') {
+    propSchema.aetherType = 'AetherDate'
+    propSchema.type = 'date'
+  } else if (propEnum) {
+    propSchema.aetherType = 'AetherEnum'
+    propSchema.type = 'enums'
+    propSchema.schema = propEnum.reduce(
+      (acc, enumVal) => ({
+        ...(acc as Record<string, string | number>),
+        [enumVal as string]: enumVal,
+      }),
+      {}
+    )
+  } else if (propType === 'string' && description?.toLowerCase().includes('color')) {
+    // TODO: Figure out a better way to detect color fields
+    propSchema.aetherType = 'AetherColor'
+    propSchema.type = 'string'
+  } else {
+    propSchema.aetherType = SCHEMA_MAP[propType!]
+    propSchema.type = propType
+  }
+  // Return
+  return propSchema
+}
+
+/** --- parseProp() ---------------------------------------------------------------------------- */
+/** -i- Convert a JSON schema prop definition to an aetherspace prop definition */
+const parseProp = (propKey: string, propDef: JSONSchema7, parseContext: { required: string[] }) => {
+  // @ts-ignore
+  const { type: propType, default: propDefault, description } = propDef
+  const required = parseContext?.required
+  let propSchema = {} as Record<string, unknown>
+  // Determine if optional
+  propSchema.isOptional = required ? !required.includes(propKey) : true
+  // Determine types
+  propSchema = parseType(propDef, propSchema)
+  // Handle descriptions
+  if (description && propSchema.type !== 'object') propSchema.description = description
+  // Handle defaults
+  if (propDefault) propSchema.defaultValue = propDefault
+  // Handle objects
+  if (propType === 'object') {
+    propSchema = { ...propSchema, ...parseSchema(propDef as JSONSchema7) }
+  }
+  // Handle arrays
+  if (propSchema.aetherType === 'AetherArray') {
+    propSchema.schema = parseProp(propKey, propDef.items as Record<string, unknown>, parseContext)
+  }
+  // Return parsed schema
+  return propSchema
+}
+
+/** --- parseSchema() -------------------------------------------------------------------------- */
+/** -i- Convert a JSON schema object to an aetherspace compatible schema */
+const parseSchema = (schema: JSONSchema7) => {
+  // @ts-ignore
+  const { type, description, properties, required = [] } = schema
+  // Do nothing when not an object
+  if (type !== 'object') return {}
+  // Build aether schema
+  const resultSchema = {} as Record<string, any>
+  resultSchema.aetherType = 'AetherSchema'
+  resultSchema.type = type
+  resultSchema.schemaName = description
+  // Parse properties
+  resultSchema.schema = Object.entries(properties!).reduce((atSchema, [propKey, propDef]) => {
+    const propSchema = parseProp(propKey, propDef as Record<string, unknown>, { required })
+    // Abort if unknown
+    if (!propSchema.aetherType) {
+      console.warn(`Unknown prop type: ${propKey} (${propSchema.type})`, propSchema)
+      return atSchema
     }
-    // Keep existing value if present
-    if (obj?.[name]) return obj[name]
-    // Apply example value instead? (docs)
-    if (applyExamples) return { ...acc, [name]: fieldConfig.example || fieldConfig.defaultValue } // prettier-ignore
-    // Apply default value
-    return { ...acc, [name]: fieldConfig.defaultValue }
+    // Return prop definition
+    return { ...atSchema, [propKey]: propSchema }
   }, {})
-  // Build final object
-  return { ...obj, ...objWithDefaults }
+  // Return parsed schema
+  return resultSchema as AetherSchemaType
 }
 
-/* --- Exports --------------------------------------------------------------------------------- */
+/* --- Object Schemas -------------------------------------------------------------------------- */
 
-export const AetherSchemaTypes = {
-  // -- Primitives --
-  string: AetherString,
-  number: AetherNumber,
-  boolean: AetherBoolean,
-  // -- Single values --
-  id: AetherID,
-  color: AetherColor,
-  date: AetherDate,
-  enum: AetherEnum,
-  // -- Objectlikes --
-  schema: AetherSchema,
-  object: AetherSchema,
-  // -- Arraylikes --
-  array: AetherArray,
-  collection: AetherCollection,
-  // -- Utilities --
-  extend: extendSchema,
-  assign: extendSchema,
-  omit: omitSchemaProps,
-  pick: pickSchemaProps,
-  partial: makePartialSchema,
+if (!z.ZodObject.prototype.aetherType) {
+  z.ZodObject.prototype.aetherType = 'AetherObject'
+  z.ZodObject.prototype.nameSchema = function (schemaName: string) {
+    this.aetherType = 'AetherSchema'
+    this.schemaName = schemaName
+    return this.describe(schemaName)
+  }
+  // Allow named extensions
+  z.ZodObject.prototype.extendSchema = function (
+    schemaName: string,
+    extraProperties: z.ZodRawShape
+  ) {
+    const extendedSchema = this.extend(extraProperties).nameSchema(schemaName)
+    if (this.exampleValue) extendedSchema.example(this.exampleValue)
+    if (this.schemaName) extendedSchema.extendedFrom = this.schemaName
+    return extendedSchema
+  }
+  // @ts-ignore
+  z.ZodObject.prototype.pickSchema = function (schemaName: string, picks: any) {
+    const pickedSchema = this.pick(picks).nameSchema(schemaName)
+    if (this.exampleValue) pickedSchema.example(this.exampleValue)
+    if (this.schemaName) pickedSchema.extendedFrom = this.schemaName
+    return pickedSchema
+  }
+  // @ts-ignore
+  z.ZodObject.prototype.omitSchema = function (schemaName: string, omits: any) {
+    const omittedSchema = this.omit(omits).nameSchema(schemaName)
+    if (this.exampleValue) omittedSchema.example(this.exampleValue)
+    if (this.schemaName) omittedSchema.extendedFrom = this.schemaName
+    return omittedSchema
+  }
+  // @ts-ignore
+  z.ZodObject.prototype.partialSchema = function (schemaName: string) {
+    const partialSchema = this.partial().nameSchema(schemaName)
+    if (this.exampleValue) partialSchema.example(this.exampleValue)
+    if (this.schemaName) partialSchema.extendedFrom = this.schemaName
+    return partialSchema
+  }
+  // @ts-ignore
+  z.ZodObject.prototype.deepPartialSchema = function (schemaName: string) {
+    const partialSchema = this.deepPartial().nameSchema(schemaName)
+    if (this.exampleValue) partialSchema.example(this.exampleValue)
+    if (this.schemaName) partialSchema.extendedFrom = this.schemaName
+    return partialSchema
+  }
+  // @ts-ignore
+  z.ZodObject.prototype.requiredSchema = function (schemaName: string) {
+    const requiredSchema = this.required().nameSchema(schemaName)
+    if (this.exampleValue) requiredSchema.example(this.exampleValue)
+    if (this.schemaName) requiredSchema.extendedFrom = this.schemaName
+    return requiredSchema
+  }
+  // Documentation
+  z.ZodObject.prototype.jsonSchema = function () {
+    return zodToJsonSchema(this) as JSONSchema7
+  }
+  z.ZodObject.prototype.introspect = function () {
+    return parseSchema(this.jsonSchema())
+  }
+  // Allow examples
+  z.ZodObject.prototype.example = function (value: any) {
+    this.exampleValue = value
+    return this
+  }
+  z.ZodObject.prototype.eg = z.ZodObject.prototype.example
+  z.ZodObject.prototype.ex = z.ZodObject.prototype.example
 }
 
-export const ats = {
-  ...AetherSchemaTypes,
-  is: ss.is,
-  validate: ss.validate,
-  assert: ss.assert,
-  applySchema,
+/** --- aetherSchema() ------------------------------------------------------------------------- */
+/** -i- Define your datastructure (Props, Args, Models) as a zod powered single source of truth */
+export const aetherSchema = <T extends z.ZodRawShape>(schemaName: string, shape: T) => {
+  return z.object(shape).nameSchema(schemaName)
 }
 
-export default ats
+/** --- buildSchema() -------------------------------------------------------------------------- */
+/** -i- Define your datastructure (Props, Args, Models) as a zod object and turn it into an aetherSchema */
+export const buildSchema = <T extends z.ZodRawShape>(
+  schemaName: string,
+  zodSchema: z.ZodObject<T>
+) => {
+  return aetherSchema(schemaName, zodSchema.shape)
+}
