+++ packages/@aetherspace/schemas/index.ts
@@ -1,4 +1,3 @@
-export * from './zodExtensions'
 export * from './aetherSchemas'
 export * from './aetherSchemaPlugin'
 export * from './aetherStoryDocs'
