+++ packages/@aetherspace/scripts/schema-test.ts
@@ -1,64 +1,5 @@
-import { ats, AetherSchemaType, aetherSchema } from '../schemas'
-// import { aetherSchema, makeSchemaDeepPartial, makeSchemaPartial, makeSchemaRequired } from '../schemas/zodSchemas'
+import { aetherSchema } from '../schemas'
 import { z } from 'zod'
-
-enum TEST_ENUM {
-    A = 'A',
-    B = 'B',
-}
-
-const id = ats.id().optional().default('some-id')
-const ids = ats.array(ats.id())
-const str = ats.string().nullable().docs('example', 'description')
-const day = ats.date().optional().docs('01/01/2022', 'The start of the year')
-const num = ats.number().docs(5)
-const bln = ats.boolean().optional()
-const opt = ats.enum<TEST_ENUM>(Object.values(TEST_ENUM))
-const obj = ats.object('StringObject', { str })
-const col = ats.array(ats.object('IDObject', { id }))
-const coll = ats.collection('IDObject', { id })
-
-// -- Build Original Schema --
-
-const originalSchema = ats.schema('MySchema', {
-    id,
-    ids,
-    str,
-    day,
-    num,
-    bln,
-    opt,
-    obj,
-    col,
-    // coll,
-})
-type OriginalSchemaType = typeof originalSchema['TYPE'] // Hover to preview
-
-// -- Test Utilities --
-
-const omittedSchema = ats.omit('Minimal', originalSchema, ['obj', 'col'])
-type OmittedSchemaType = typeof omittedSchema['TYPE']
-
-const extendedSchema = ats.extend('Extended', omittedSchema, {
-    extra: ats.string().nullish(),
-})
-type ExtendedSchemaType = typeof extendedSchema['TYPE'] // Hover to preview
-
-const partialSchema = ats.partial('Partial', extendedSchema)
-type PartialSchemaType = typeof partialSchema['TYPE'] // Hover to preview
-
-const pickedSchema = ats.pick('Picked', partialSchema, ['id', 'str', 'day'])
-type PickedSchemaType = typeof pickedSchema['TYPE'] // Hover to preview
-
-// -- Test Results --
-
-const finalSchema = originalSchema // pickedSchema // partialSchema // extendedSchema // omittedSchema // originalSchema
-type FinalSchemaType = typeof finalSchema['TYPE'] // Hover to preview
-
-console.log('--- ATS ---\n', JSON.stringify(finalSchema, null, 4))
-// ↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓↓
-
-// /* --- Zod ------------------------------------------------------------------------------------- */
 
 const Primitives = aetherSchema('Primitives', {
     bln: z.boolean().optional().describe('A boolean'),
